Google style guide

+++ service/src/main/java/com/codeforcommunity/processor/ProtectedUserProcessorImpl.java
@@ -1,5 +1,7 @@
 package com.codeforcommunity.processor;
 
+import static org.jooq.generated.Tables.*;
+
 import com.codeforcommunity.api.IProtectedUserProcessor;
 import com.codeforcommunity.auth.JWTData;
 import com.codeforcommunity.auth.Passwords;
@@ -9,14 +11,11 @@ import com.codeforcommunity.dto.protected_user.components.Contact;
 import com.codeforcommunity.dto.user.ChangePasswordRequest;
 import com.codeforcommunity.exceptions.UserDoesNotExistException;
 import com.codeforcommunity.exceptions.WrongPasswordException;
+import java.util.List;
 import org.jooq.DSLContext;
 import org.jooq.generated.tables.records.ChildrenRecord;
 import org.jooq.generated.tables.records.ContactsRecord;
 import org.jooq.generated.tables.records.UsersRecord;
-
-import java.util.List;
-
-import static org.jooq.generated.Tables.*;
 
 public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
 
@@ -30,34 +29,25 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
   public void deleteUser(JWTData userData) {
     int userId = userData.getUserId();
 
-    db.deleteFrom(EVENT_REGISTRATIONS)
-        .where(EVENT_REGISTRATIONS.USER_ID.eq(userId))
-        .executeAsync();
+    db.deleteFrom(EVENT_REGISTRATIONS).where(EVENT_REGISTRATIONS.USER_ID.eq(userId)).executeAsync();
 
-    db.deleteFrom(VERIFICATION_KEYS)
-        .where(VERIFICATION_KEYS.USER_ID.eq(userId))
-        .executeAsync();
+    db.deleteFrom(VERIFICATION_KEYS).where(VERIFICATION_KEYS.USER_ID.eq(userId)).executeAsync();
 
-    db.deleteFrom(PF_REQUESTS)
-        .where(PF_REQUESTS.USER_ID.eq(userId))
-        .executeAsync();
+    db.deleteFrom(PF_REQUESTS).where(PF_REQUESTS.USER_ID.eq(userId)).executeAsync();
 
-    db.deleteFrom(USERS)
-        .where(USERS.ID.eq(userId))
-        .executeAsync();
+    db.deleteFrom(USERS).where(USERS.ID.eq(userId)).executeAsync();
   }
 
   @Override
   public void changePassword(JWTData userData, ChangePasswordRequest changePasswordRequest) {
-    UsersRecord user = db.selectFrom(USERS)
-        .where(USERS.ID.eq(userData.getUserId()))
-        .fetchOne();
+    UsersRecord user = db.selectFrom(USERS).where(USERS.ID.eq(userData.getUserId())).fetchOne();
 
     if (user == null) {
       throw new UserDoesNotExistException(userData.getUserId());
     }
 
-    if (Passwords.isExpectedPassword(changePasswordRequest.getCurrentPassword(), user.getPassHash())) {
+    if (Passwords.isExpectedPassword(
+        changePasswordRequest.getCurrentPassword(), user.getPassHash())) {
       user.setPassHash(Passwords.createHash(changePasswordRequest.getNewPassword()));
       user.store();
     } else {
@@ -66,27 +56,27 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
   }
 
   @Override
-  public void setContactsAndChildren(JWTData userData, SetContactsAndChildrenRequest setContactsAndChildrenRequest) {
+  public void setContactsAndChildren(
+      JWTData userData, SetContactsAndChildrenRequest setContactsAndChildrenRequest) {
 
     updateMainContact(userData, setContactsAndChildrenRequest.getMainContact());
 
-    if (setContactsAndChildrenRequest.getChildren() != null &&
-            !setContactsAndChildrenRequest.getChildren().isEmpty()) {
+    if (setContactsAndChildrenRequest.getChildren() != null
+        && !setContactsAndChildrenRequest.getChildren().isEmpty()) {
       addChildren(setContactsAndChildrenRequest.getChildren(), userData);
     }
 
-    if (setContactsAndChildrenRequest.getAdditionalContacts() != null &&
-            !setContactsAndChildrenRequest.getAdditionalContacts().isEmpty()) {
+    if (setContactsAndChildrenRequest.getAdditionalContacts() != null
+        && !setContactsAndChildrenRequest.getAdditionalContacts().isEmpty()) {
       addAdditionalContacts(setContactsAndChildrenRequest.getAdditionalContacts(), userData);
     }
-
   }
-  
+
   private void updateMainContact(JWTData userData, Contact newContactData) {
 
-    ContactsRecord mainContact = db.selectFrom(CONTACTS)
-            .where(CONTACTS.USER_ID.eq(userData.getUserId())
-                    .and(CONTACTS.IS_MAIN_CONTACT))
+    ContactsRecord mainContact =
+        db.selectFrom(CONTACTS)
+            .where(CONTACTS.USER_ID.eq(userData.getUserId()).and(CONTACTS.IS_MAIN_CONTACT))
             .fetchOne();
 
     if (mainContact == null) {
@@ -106,9 +96,8 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
 
     mainContact.store();
 
-    UsersRecord usersRecord = db.selectFrom(USERS)
-            .where(USERS.ID.eq(userData.getUserId()))
-                    .fetchOne();
+    UsersRecord usersRecord =
+        db.selectFrom(USERS).where(USERS.ID.eq(userData.getUserId())).fetchOne();
 
     usersRecord.setEmail(newContactData.getEmail());
 
@@ -116,7 +105,7 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
   }
 
   private void addChildren(List<Child> children, JWTData userData) {
-    
+
     for (Child c : children) {
 
       ChildrenRecord childrenRecord = db.newRecord(CHILDREN);
@@ -131,13 +120,13 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
       childrenRecord.setDiagnosis(c.getDiagnosis());
       childrenRecord.setMedications(c.getMedications());
       childrenRecord.setNotes(c.getNotes());
-      
+
       childrenRecord.store();
     }
   }
 
   private void addAdditionalContacts(List<Contact> additionalContacts, JWTData userData) {
-    
+
     for (Contact c : additionalContacts) {
 
       ContactsRecord contactsRecord = db.newRecord(CONTACTS);
@@ -156,6 +145,5 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
 
       contactsRecord.store();
     }
-  } 
-  
+  }
 }
